+++ src/main/kotlin/com/pestphp/pest/PestTestRunLineMarkerProvider.kt
@@ -4,7 +4,6 @@ import com.intellij.execution.lineMarker.RunLineMarkerContributor
 import com.intellij.psi.PsiElement
 import com.jetbrains.php.lang.lexer.PhpTokenTypes
 import com.jetbrains.php.lang.psi.PhpPsiUtil
-import com.jetbrains.php.lang.psi.elements.FunctionReference
 import com.jetbrains.php.lang.psi.elements.impl.FunctionReferenceImpl
 
 class PestTestRunLineMarkerProvider : RunLineMarkerContributor() {
